Handle create issue command with unknown issue type
When the issue type is unknown, the API returns an error with the
`issuetype` key. In that case, fetch a list of non-subtask types and
display to the user the valid types.

Also, creating epics requires some other special info when creating the
ticket, so specify that we can't create epics yet.